Move handlers to versioned namespace
