+++ consoleme/handlers/v2/__init__.py
@@ -1,0 +1,1 @@
+"""Endpoint handlers."""
