+++ consoleme/routes.py
@@ -13,19 +13,19 @@ from apispec_webframeworks.tornado import TornadoPlugin
 from raven.contrib.tornado import AsyncSentryClient
 
 from consoleme.config import config
-from consoleme.handlers.autologin import AutoLoginHandler
 from consoleme.handlers.base import NoCacheStaticFileHandler
-from consoleme.handlers.credentials import GetCredentialsHandler
-from consoleme.handlers.dynamic_config import DynamicConfigHandler
-from consoleme.handlers.errors import Consolme404Handler
-from consoleme.handlers.headers import (
+from consoleme.handlers.v1.autologin import AutoLoginHandler
+from consoleme.handlers.v1.credentials import GetCredentialsHandler
+from consoleme.handlers.v1.dynamic_config import DynamicConfigHandler
+from consoleme.handlers.v1.errors import Consolme404Handler
+from consoleme.handlers.v1.headers import (
     ApiHeaderHandler,
     HeaderHandler,
     PageHeaderHandler,
 )
-from consoleme.handlers.health import HealthHandler
-from consoleme.handlers.index import IndexHandler
-from consoleme.handlers.policies import (
+from consoleme.handlers.v1.health import HealthHandler
+from consoleme.handlers.v1.index import IndexHandler
+from consoleme.handlers.v1.policies import (
     AutocompleteHandler,
     GetPoliciesHandler,
     PolicyEditHandler,
@@ -38,10 +38,10 @@ from consoleme.handlers.policies import (
     SelfServiceHandler,
 )
 
-from consoleme.handlers.roles import GetRolesHandler
-from consoleme.handlers.saml import SamlHandler
+from consoleme.handlers.v1.roles import GetRolesHandler
+from consoleme.handlers.v1.saml import SamlHandler
 
-from consoleme.handlers.swagger import SwaggerHandler, SwaggerJsonGenerator
+from consoleme.handlers.v1.swagger import SwaggerHandler, SwaggerJsonGenerator
 
 from consoleme.lib.auth import mk_jwks_validator
 from consoleme.lib.plugins import get_plugin_by_name
